feat: 기본 요구사항 및 심화 요구 사항구현

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicChannelService.java
@@ -127,6 +127,9 @@ public class BasicChannelService implements ChannelService {
 
     @Override
     public void deleteById(UUID id) {
+        if (channelRepository.findById(id) == null) {
+            throw new NotFoundException("존재하지 않는 채널입니다.");
+        }
         MessageDeletionSupport.deleteByChannelId(messageRepository, binaryContentRepository, id);
         // 관련 ReadStatus 삭제
         readStatusRepository.deleteByChannelId(id);
